refactor(orders-table): rename OrdersWidget to OrdersTableWidget
